Refactor and fix component naming
