--- src/components/ProposalItemBody.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts"></script>
-
-<template>
-  <p class="mb-2 break-words text-md line-clamp-2">
-    <slot />
-  </p>
-</template>
