+++ src/components/SpaceProposalResultsProgressBar.vue
