+++ src/components/SpaceProposalResultsQuorumPlugin.vue
