+++ src/components/SpaceProposalResultsShutter.vue
